Reference

In this reference you will find a detailed overview of the package API.

Reference guides are technical descriptions of the machinery and how to operate it. Reference material is information-oriented.

–- Diátaxis

In other words, you come here because you want to take a very close look at the code 🧐

Content

Reference

Content

Index

Public Interface

Internal functions

Index

AlgorithmicRecourseDynamics.run!

AlgorithmicRecourseDynamics.run_experiment

AlgorithmicRecourseDynamics.run_experiment

AlgorithmicRecourseDynamics.run_experiments

AlgorithmicRecourseDynamics.run_experiments

AlgorithmicRecourseDynamics.set_up_experiment

AlgorithmicRecourseDynamics.set_up_experiment

AlgorithmicRecourseDynamics.set_up_experiments

Public Interface

AlgorithmicRecourseDynamics.run_experiment — Method

run_experiment(
    experiment::Experiment; evaluate_every::Int=2,
    save_path::Union{Nothing,String}=nothing,
    save_name::Union{Nothing,String}=nothing,
    kwargs...
)

Runs a given experiment and saves the results if specified.

AlgorithmicRecourseDynamics.run_experiment — Method

run_experiment(
    data::CounterfactualData,
    models::Dict{Symbol, <: CounterfactualExplanations.Models.AbstractFittedModel},
    generators::Dict{Symbol, <: CounterfactualExplanations.Generators.AbstractGenerator};
    target::Int=1,
    num_counterfactuals::Int=5,
    evaluate_every::Int=2,
    pre_train_models::Union{Nothing,Int}=100,
    save_path::Union{Nothing,String}=nothing,
    save_name::Union{Nothing,String}=nothing,
    kwargs...
)

Sets up one experiment for the provided data, models and generators and then runs it. Saves results if specified. Models and generators need to be supplied as dictionaries, where values need to be of type CounterfactualExplanations.Models.AbstractFittedModel and CounterfactualExplanations.Generators.AbstractGenerator, respectively.

AlgorithmicRecourseDynamics.run_experiments — Method

function run_experiments(
    experiments::Dict{Symbol, Experiment};
    evaluate_every::Int=2,
    save_path::Union{Nothing,String}=nothing,
    kwargs...
)

Runs multiple provided experiments.

AlgorithmicRecourseDynamics.run_experiments — Method

run_experiments(
    catalogue::Dict{Symbol, CounterfactualData},
    models::Union{Dict{Symbol, <: CounterfactualExplanations.Models.AbstractFittedModel},Vector{Symbol}},
    generators::Dict{Symbol, <: CounterfactualExplanations.Generators.AbstractGenerator};
    target::Int=1,
    num_counterfactuals::Int=5,
    evaluate_every::Int=2,
    pre_train_models::Union{Nothing, Int}=100,
    save_path::Union{Nothing,String}=nothing,
    kwargs...
)

Sets up and runs experiments for multiple data sets.

AlgorithmicRecourseDynamics.set_up_experiment — Method

set_up_experiment(
    data::CounterfactualData,
    models::Dict{Symbol, <: CounterfactualExplanations.Models.AbstractFittedModel},
    generators::Dict{Symbol, <: CounterfactualExplanations.Generators.AbstractGenerator};
    target::Int=1,
    num_counterfactuals::Int=5,
    pre_train_models::Union{Nothing,Int}=100,
    kwargs...
)

Sets up one experiment for the provided data, models and generators.

AlgorithmicRecourseDynamics.set_up_experiment — Method

set_up_experiment(
    data::CounterfactualData,
    models::Dict{Symbol, <: CounterfactualExplanations.Models.AbstractFittedModel},
    generators::Dict{Symbol, <: CounterfactualExplanations.Generators.AbstractGenerator};
    target::Int=1,
    num_counterfactuals::Int=5,
    pre_train_models::Union{Nothing,Int}=100,
    kwargs...
)

Sets up one experiment for the provided data, models and generators.

AlgorithmicRecourseDynamics.set_up_experiments — Method

function set_up_experiments(
    catalogue::Dict{Symbol, CounterfactualData},
    models::Union{Dict{Symbol, <: CounterfactualExplanations.Models.AbstractFittedModel},Vector{Symbol}},
    generators::Dict{Symbol, <: CounterfactualExplanations.Generators.AbstractGenerator};
    target::Int=1,
    num_counterfactuals::Int=5,
    pre_train_models::Union{Nothing, Int}=100,
    kwargs...
)

Sets up multiple experiments.

Internal functions

AlgorithmicRecourseDynamics.run! — Method

run!(experiment::Experiment, generator::CounterfactualExplanations.AbstractGenerator, n_folds=5; seed=nothing, T=1000)

A wrapper function that runs the experiment for endogenous models shifts.